+++ hooks/useOrg.ts
@@ -1,17 +1,28 @@
+// /organizations/hooks/useOrganizations.ts
+/* eslint-disable @typescript-eslint/no-explicit-any */
+"use client";
+
 import { useState, useEffect, useMemo, useCallback } from "react";
 import { toast } from "sonner";
+import axios from "axios";
 import { Organization } from "../app/(dashboard)/super-admin/(pages)/organizations/org";
 import { useAuth } from "@/components/auth";
 import { isTokenExpired } from "@/hooks/token";
-import axios from "axios";
 
 export function useOrganizations() {
-  const { token, refreshAccessToken } = useAuth();
   const [orgs, setOrgs] = useState<Organization[]>([]);
   const [loading, setLoading] = useState(true);
+  const { token, refreshAccessToken } = useAuth();
   const [searchTerm, setSearchTerm] = useState("");
   const [statusFilter, setStatusFilter] = useState("all");
+  const [viewOpen, setViewOpen] = useState(false);
+  const [editOpen, setEditOpen] = useState(false);
+  const [selectedOrg, setSelectedOrg] = useState<Organization | null>(null);
+  const [editFormData, setEditFormData] = useState<Organization>(
+    {} as Organization,
+  );
 
+  // Fetch organizations
   useEffect(() => {
     if (!token) return;
     let isMounted = true;
@@ -33,7 +44,20 @@ export function useOrganizations() {
         );
 
         if (!isMounted) return;
-        const mapped = res.data.map((o: any) => ({
+
+        type ApiOrg = {
+          id: string;
+          name: string;
+          invite_code: string;
+          ico_registered: boolean;
+          data_retention_years: number;
+          type: string;
+          is_active: boolean;
+          created_at: string;
+          user_count: number;
+        };
+
+        const mapped: Organization[] = (res.data as ApiOrg[]).map((o) => ({
           id: o.id,
           name: o.name,
           type: o.type,
@@ -44,9 +68,13 @@ export function useOrganizations() {
           ico_registered: o.ico_registered,
           data_retention_years: o.data_retention_years,
         }));
+
         setOrgs(mapped);
-      } catch (err: any) {
-        toast.error(err?.response?.data?.detail || err.message);
+      } catch (e: any) {
+        console.error(
+          "Failed to load organizations:",
+          e?.response?.data?.detail || e.message,
+        );
       } finally {
         if (isMounted) setLoading(false);
       }
@@ -56,24 +84,276 @@ export function useOrganizations() {
     return () => {
       isMounted = false;
     };
-  }, [token]);
+  }, [token, refreshAccessToken]);
 
-  const filteredOrganizations = useMemo(() => {
-    return orgs.filter(
-      (org) =>
+  // Filtered orgs
+  const filteredOrgs = useMemo(() => {
+    return orgs.filter((org) => {
+      const matchesSearch =
         org.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
-        org.type.toLowerCase().includes(searchTerm.toLowerCase()),
-    );
+        org.type.toLowerCase().includes(searchTerm.toLowerCase());
+
+      if (statusFilter === "all") return matchesSearch;
+      return (
+        matchesSearch &&
+        org.status?.toString().toLowerCase() === statusFilter.toLowerCase()
+      );
+    });
   }, [orgs, searchTerm, statusFilter]);
+
+  // Create org
+  const createOrg = async (orgData: any) => {
+    try {
+      let validToken = token;
+      if (!validToken || isTokenExpired(validToken)) {
+        const refreshed = await refreshAccessToken();
+        if (!refreshed) {
+          toast.error("Authentication expired. Please log in again.");
+          return;
+        }
+        validToken = refreshed;
+      }
+
+      const res = await axios.post(
+        "https://medilogic-backend.onrender.com/super/organizations",
+        orgData,
+        {
+          headers: {
+            Authorization: `Bearer ${validToken}`,
+            "Content-Type": "application/json",
+          },
+        },
+      );
+
+      setOrgs((prev) => [
+        ...prev,
+        {
+          id: res.data.id,
+          name: res.data.name,
+          type: res.data.type,
+          status: res.data.is_active,
+          userCount: res.data.user_count ?? 0,
+          createdDate: new Date(res.data.created_at).toLocaleDateString(),
+          invite_code: res.data.invite_code,
+          ico_registered: res.data.ico_registered,
+          data_retention_years: res.data.data_retention_years,
+        },
+      ]);
+
+      toast.success("Organization created successfully");
+    } catch (err: any) {
+      const detail = err?.response?.data?.detail;
+      const msg = Array.isArray(detail)
+        ? detail.map((d: any) => d.msg).join(" • ")
+        : detail || err.message || "Failed to create organization";
+      toast.error(msg);
+      throw err;
+    }
+  };
+
+  // View org
+  const viewOrg = async (org: Organization) => {
+    const orgId = org.id;
+    const existingOrg = orgs.find((o) => o.id === orgId);
+
+    if (existingOrg && existingOrg.description) {
+      setSelectedOrg(existingOrg);
+      setViewOpen(true);
+      return;
+    }
+
+    try {
+      let validToken = token;
+      if (!validToken || isTokenExpired(validToken)) {
+        const refreshed = await refreshAccessToken();
+        if (!refreshed) {
+          toast.error("Authentication expired. Please log in again.");
+          return;
+        }
+        validToken = refreshed;
+      }
+
+      const res = await axios.get(
+        `https://medilogic-backend.onrender.com/super/${orgId}`,
+        { headers: { Authorization: `Bearer ${validToken}` } },
+      );
+
+      const data = res.data;
+      const mappedOrg: Organization = {
+        id: data.organization.id,
+        name: data.organization.name,
+        type: data.organization.type ?? "",
+        status: data.organization.is_active,
+        createdDate: new Date(
+          data.organization.created_at,
+        ).toLocaleDateString(),
+        userCount: data.user_count ?? 0,
+        description: data.organization.description ?? "",
+        address: data.organization.address ?? "",
+        phone: data.organization.phone ?? "",
+        email: data.organization.email ?? "",
+        invite_code: data.organization.invite_code,
+        ico_registered: data.organization.ico_registered,
+        data_retention_years: data.organization.data_retention_years,
+      };
+
+      setSelectedOrg(mappedOrg);
+      setViewOpen(true);
+    } catch {
+      toast.error("Failed to load organization details");
+    }
+  };
+
+  // Edit org
+  const editOrg = async (org: Organization) => {
+    setSelectedOrg(org);
+    try {
+      let validToken = token;
+      if (!validToken || isTokenExpired(validToken)) {
+        const refreshed = await refreshAccessToken();
+        if (!refreshed) return;
+        validToken = refreshed;
+      }
+
+      const res = await axios.get(
+        `https://medilogic-backend.onrender.com/super/${org.id}`,
+        { headers: { Authorization: `Bearer ${validToken}` } },
+      );
+
+      const data = res.data;
+      setEditFormData({
+        id: data.organization.id,
+        name: data.organization.name,
+        type: data.organization.type,
+        description: data.organization.description ?? "",
+        address: data.organization.address_line ?? "",
+        phone: data.organization.phone_number ?? "",
+        email: data.organization.email ?? "",
+        postal_code: data.organization.postal_code ?? "",
+        license_number: data.organization.license_number ?? "",
+        waste_processing_capability:
+          data.organization.waste_processing_capability ?? "",
+        delivery_capacity: data.organization.delivery_capacity ?? 0,
+        contact_person_name: data.organization.contact_person_name ?? "",
+        contact_person_role: data.organization.contact_person_role ?? "",
+        latitude: data.organization.latitude ?? 0,
+        longitude: data.organization.longitude ?? 0,
+        status: data.organization.is_active,
+        createdDate: data.organization.createDate,
+        userCount: data.organization.user_count,
+      });
+      setEditOpen(true);
+    } catch {
+      toast.error("Failed to load organization details for editing");
+    }
+  };
+
+  const handleEditChange = (changes: Partial<Organization>) => {
+    setEditFormData((prev: any) => ({ ...prev, ...changes }));
+  };
+
+  const handleSaveEdit = async () => {
+    if (!selectedOrg || !editFormData) return;
+
+    try {
+      let validToken = token;
+      if (!validToken || isTokenExpired(validToken)) {
+        const refreshed = await refreshAccessToken();
+        if (!refreshed) {
+          toast.error("Authentication expired. Please log in again.");
+          return;
+        }
+        validToken = refreshed;
+      }
+
+      const payload = {
+        id: selectedOrg.id,
+        name: editFormData.name,
+        type: editFormData.type,
+        description: editFormData.description,
+        address_line: editFormData.address,
+        phone_number: editFormData.phone,
+        email: editFormData.email,
+        postal_code: editFormData.postal_code ?? "",
+        license_number: editFormData.license_number ?? "",
+        waste_processing_capability:
+          editFormData.waste_processing_capability ?? "",
+        delivery_capacity: editFormData.delivery_capacity ?? "",
+        contact_person_name: editFormData.contact_person_name ?? "",
+        contact_person_role: editFormData.contact_person_role ?? "",
+        latitude: editFormData.latitude ?? 0,
+        longitude: editFormData.longitude ?? 0,
+        is_active: editFormData.status ?? true,
+      };
+
+      const res = await axios.patch(
+        `https://medilogic-backend.onrender.com/super/${selectedOrg.id}`,
+        payload,
+        {
+          headers: {
+            Authorization: `Bearer ${validToken}`,
+            "Content-Type": "application/json",
+          },
+        },
+      );
+
+      setOrgs((prev) =>
+        prev.map((org) =>
+          org.id === selectedOrg.id ? { ...org, ...res.data } : org,
+        ),
+      );
+
+      toast.success(`${payload.name} has been updated successfully`);
+      setEditOpen(false);
+      setSelectedOrg(null);
+    } catch (err: any) {
+      const msg = err?.response?.data
+        ? JSON.stringify(err.response.data)
+        : err.message || "Failed to update organization";
+      toast.error(msg);
+    }
+  };
+
+  // Misc actions
+  const deactivateOrg = useCallback((orgName: string) => {
+    toast.success(`${orgName} has been deactivated`);
+  }, []);
+
+  const regenerateInviteCode = useCallback((orgName: string) => {
+    toast.success(`Invite code regenerated for ${orgName}`);
+  }, []);
+
+  // Close dialogs
+  const closeView = () => {
+    setViewOpen(false);
+    setSelectedOrg(null);
+  };
+
+  const closeEdit = () => {
+    setEditOpen(false);
+    setSelectedOrg(null);
+  };
 
   return {
     orgs,
-    setOrgs,
+    filteredOrgs,
     loading,
-    filteredOrganizations,
     searchTerm,
     setSearchTerm,
     statusFilter,
     setStatusFilter,
+    viewOpen,
+    editOpen,
+    selectedOrg,
+    editFormData,
+    createOrg,
+    viewOrg,
+    editOrg,
+    handleEditChange,
+    handleSaveEdit,
+    deactivateOrg,
+    regenerateInviteCode,
+    closeView,
+    closeEdit,
   };
 }
